refactor: move BaseButton to folder
